3D Models
I began the process of renaming the tetrix parts, still have a little
ways to go.  I also made an assembly of a simple squarebot for an
example, as well as a rolling wheel assembly that both teams can copy
into their models instaed of assembling four on their own (not fun i
tried) As for PHI Alpha, i began to model a possible chassis design that
we will look at on monday.